+++ HISTORY.md
@@ -5,16 +5,20 @@
 ### Breaking Behavior
 
 * Negative values no longer supported in `region`  in xarray `from_xarray` and `copy_data_from_xarray`.
+* Removes VirtualGroup class.
+* Removes Group class.
 
 ### Deprecation
 
-* Removes virtual groups.
+### Deprecation
 
 ### Bug Fixes
 
 ### New Features
 
 * Adds `skip_vars` argument to the methods for copying xarray Dataset to TileDB that allows the user to specify variables to skip.
+* Adds method `create_group` that creates a TileDB group and arrays.
+* Adds method `open_group_array` that opens a TileDB array from a provided Group.
 
 ### Improvements